requirements.txt
asgiref==3.5.1
Django==4.0.4
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2

requests~=2.27.1
openpyxl~=3.0.9
beautifulsoup4~=4.11.1